User 모델 재설정

+++ layerd/repositories/likes.repository.js
@@ -1,10 +1,19 @@
+const { UnexpectedError } = require('../../middlewares/custom-exception');
+
 class LikesRepository {
-  constructor(likesModel) {
+  constructor(likesModel, postModel) {
     this.likesModel = likesModel;
+    this.postModel = postModel;
   }
 
   likePost = async (postId, userId) => {
+    const existPost = await this.postModel.findOne({ where: { postId } });
+
     const isLike = await this.likesModel.findAll({ where: { postId, userId } });
+    if (!existPost) {
+      console.log(2);
+      throw new UnexpectedError('없는 게시글입니다.', 404);
+    }
     if (isLike.length) {
       await this.likesModel.destroy({ where: { postId } });
       return { message: '좋아요 취소' };
